@@ -1,5 +1,5 @@
 {
-	"name": "material-duration-picker",
+	"name": "@gvlatko/material-duration-picker-v5",
 	"description": "React Material component for durations inspired by Material-UI Pickers",
 	"keywords": [
 		"react",
@@ -10,29 +10,29 @@
 		"duration-picker",
 		"durations"
 	],
-	"version": "1.4.4",
+	"version": "2.0.0",
 	"license": "MIT",
 	"main": "dist/index.js",
-	"author": "tran-simon",
-	"repository": "https://github.com/tran-simon/material-duration-picker",
+	"author": "gvlatko",
+	"repository": "https://github.com/gvlatko/material-duration-picker",
 	"peerDependencies": {
-		"@material-ui/core": "^4.9.1",
-		"@types/react": "^16.8.4",
-		"@types/react-dom": "^16.9.0",
-		"react": "^16.8.4",
-		"react-dom": "^16.8.4"
+		"@mui/material": "^5.2.2",
+		"@types/react": "^17.0.2",
+		"@types/react-dom": "^17.0.2",
+		"react": "^17.0.2",
+		"react-dom": "^17.0.2"
 	},
 	"devDependencies": {
-		"@material-ui/core": "^4.9.1",
+		"@mui/material": "^5.2.2",
 		"@testing-library/jest-dom": "^5.12.0",
 		"@testing-library/react": "^11.2.6",
 		"@types/jest": "^26.0.23",
 		"@types/node": "^12.0.0",
 		"@types/react": "^16.8.4",
-		"@types/react-dom": "^16.9.0",
+		"@types/react-dom": "^17.0.2",
 		"jest": "^26.6.3",
-		"react": "^16.8.4",
-		"react-dom": "^16.8.4",
+		"react": "^17.0.2",
+		"react-dom": "^17.0.2",
 		"ts-jest": "^26.5.5",
 		"typescript": "^4.2.4"
 	},
@@ -45,5 +45,8 @@
 	},
 	"files": [
 		"dist"
-	]
+	],
+	"dependencies": {
+		"@emotion/styled": "^11.6.0"
+	}
 }
